refactor: move tests into contract.rs
